Require re-initialization when cache.properties is missing
Prior to this commit, the check whether initialization is required
failed when cache.properties did not exist. Since the file is only
created when `DefaultPersistentDirectoryCache` is used instead of
`DefaultPersistentDirectoryStore`, this can happen when re-using an
existing cache directory that used to be accessed using the latter but
from now on will be accessed using the former. For instance, when adding
a cleanup action, `DefaultCacheFactory` will no longer return a plain
Store but a Cache.